[refactor/#23] 회원가입 정보입력 페이지 리팩토링 (통합)
